convert types of ushort to int

+++ WinBoyEmulator/Memory/IMemory.cs
@@ -11,10 +11,10 @@ namespace WinBoyEmulator.Memory
         void Reset();
         void Load(string fileName);
 
-        byte ReadByte(ushort address);
-        ushort ReadShort(ushort address);
+        byte ReadByte(int address);
+        int ReadShort(int address);
 
-        void WriteByte(ushort address, byte value);
-        void WriteShort(ushort address, ushort value);
+        void WriteByte(int address, byte value);
+        void WriteShort(int address, int value);
     }
 }
